Move headers into public directory
