+++ src/google/encodetable.h
